refactor: #130-wasteApi, chatApi 메서드 이동 - wasteApi에서 리뷰, 관심추가 기능 따로 WasteLikeApi, WasteReviewApi로 생성 - wasteApi의 관심리스트 조회 기능 wasteLikeApi로 이동 - ChatApi의 채팅요청기능 wasteApi로 이동 - 테스트 코드도 이동
