+++ src/main/java/freshtrash/freshtrashbackend/controller/WasteApi.java
@@ -1,17 +1,18 @@
 package freshtrash.freshtrashbackend.controller;
 
 import com.querydsl.core.types.Predicate;
-import freshtrash.freshtrashbackend.dto.constants.LikeStatus;
-import freshtrash.freshtrashbackend.dto.request.ReviewRequest;
 import freshtrash.freshtrashbackend.dto.request.WasteRequest;
-import freshtrash.freshtrashbackend.dto.response.ApiResponse;
-import freshtrash.freshtrashbackend.dto.response.ReviewResponse;
+import freshtrash.freshtrashbackend.dto.response.ChatRoomResponse;
 import freshtrash.freshtrashbackend.dto.response.WasteResponse;
 import freshtrash.freshtrashbackend.dto.security.MemberPrincipal;
+import freshtrash.freshtrashbackend.entity.ChatRoom;
+import freshtrash.freshtrashbackend.entity.Member;
 import freshtrash.freshtrashbackend.entity.Waste;
 import freshtrash.freshtrashbackend.entity.constants.UserRole;
+import freshtrash.freshtrashbackend.exception.ChatRoomException;
 import freshtrash.freshtrashbackend.exception.WasteException;
 import freshtrash.freshtrashbackend.exception.constants.ErrorCode;
+import freshtrash.freshtrashbackend.service.ChatRoomService;
 import freshtrash.freshtrashbackend.service.FileService;
 import freshtrash.freshtrashbackend.service.WasteService;
 import lombok.RequiredArgsConstructor;
@@ -30,11 +31,12 @@ import javax.validation.Valid;
 import static org.springframework.data.domain.Sort.Direction.DESC;
 
 @RestController
-@RequestMapping("/api/v1/wastes")
 @RequiredArgsConstructor
+@RequestMapping("/api/v1/wastes")
 public class WasteApi {
     private final WasteService wasteService;
     private final FileService fileService;
+    private final ChatRoomService chatRoomService;
 
     /**
      * 폐기물 단일 조회
@@ -57,14 +59,6 @@ public class WasteApi {
             @QuerydslPredicate(root = Waste.class) Predicate predicate,
             @PageableDefault(size = 6, sort = "createdAt", direction = DESC) Pageable pageable) {
         Page<WasteResponse> wastes = wasteService.getWastes(district, predicate, pageable);
-        return ResponseEntity.ok(wastes);
-    }
-
-    @GetMapping("/likes")
-    public ResponseEntity<Page<WasteResponse>> getLikedWastes(
-            @PageableDefault(size = 6, sort = "createdAt", direction = DESC) Pageable pageable,
-            @AuthenticationPrincipal MemberPrincipal memberPrincipal) {
-        Page<WasteResponse> wastes = wasteService.getLikedWastes(memberPrincipal.id(), pageable);
         return ResponseEntity.ok(wastes);
     }
 
@@ -111,55 +105,36 @@ public class WasteApi {
     }
 
     /**
+     * 채팅 요청
+     */
+    @PostMapping("/{wasteId}/chats")
+    public ResponseEntity<ChatRoomResponse> handleChatRoomRequest(
+            @PathVariable Long wasteId, @AuthenticationPrincipal MemberPrincipal memberPrincipal) {
+        Waste waste = wasteService.getWaste(wasteId);
+        Member seller = waste.getMember();
+        checkIfSellerOfWaste(memberPrincipal.id(), seller.getId());
+
+        ChatRoom chatRoom = chatRoomService.getOrCreateChatRoom(seller.getId(), memberPrincipal.id(), wasteId);
+
+        ChatRoomResponse response = ChatRoomResponse.fromEntity(
+                chatRoom, waste.getTitle(), seller.getNickname(), memberPrincipal.nickname());
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    }
+
+    /**
+     * 판매자가 자신이 등록한 폐기물에 대해 채팅을 시도하는 경우 예외 처리
+     */
+    private void checkIfSellerOfWaste(Long buyerId, Long sellerId) {
+        if (sellerId.equals(buyerId)) throw new ChatRoomException(ErrorCode.CANNOT_CHAT_WITH_SELF);
+    }
+
+    /**
      * 작성자 또는 관리자가 맞는지 확인
      */
     private void checkIfWriterOrAdmin(MemberPrincipal memberPrincipal, Long wasteId) {
         if (memberPrincipal.getUserRole() != UserRole.ADMIN
                 && !wasteService.isWriterOfArticle(wasteId, memberPrincipal.id()))
             throw new WasteException(ErrorCode.FORBIDDEN_WASTE);
-    }
-
-    /**
-     * 폐기물 리뷰 작성
-     */
-    @PostMapping("/{wasteId}/reviews")
-    private ResponseEntity<ReviewResponse> addWasteReview(
-            @RequestBody @Valid ReviewRequest reviewRequest,
-            @PathVariable Long wasteId,
-            @AuthenticationPrincipal MemberPrincipal memberPrincipal) {
-
-        // TODO : transaction 거래 확인
-
-        ReviewResponse reviewResponse =
-                ReviewResponse.fromEntity(wasteService.insertWasteReview(reviewRequest, wasteId, memberPrincipal.id()));
-        return ResponseEntity.status(HttpStatus.CREATED).body(reviewResponse);
-    }
-
-    /**
-     * 폐기물 관심 추가 또는 삭제
-     */
-    @PostMapping("/{wasteId}/likes")
-    public ResponseEntity<ApiResponse<Boolean>> addOrDeleteWasteLike(
-            @RequestParam LikeStatus likeStatus,
-            @PathVariable Long wasteId,
-            @AuthenticationPrincipal MemberPrincipal memberPrincipal) {
-        checkIfNotWriter(memberPrincipal, wasteId);
-        Boolean isLike = likeStatus == LikeStatus.LIKE;
-        if (likeStatus == LikeStatus.LIKE) {
-            wasteService.addWasteLike(memberPrincipal.id(), wasteId);
-        } else if (likeStatus == LikeStatus.UNLIKE) {
-            wasteService.deleteWasteLike(memberPrincipal.id(), wasteId);
-        }
-
-        return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.of(isLike));
-    }
-
-    /**
-     * 작성자인지 확인 (작성자인 경우 관심 추가/삭제 할수 없음)
-     */
-    private void checkIfNotWriter(MemberPrincipal memberPrincipal, Long wasteId) {
-        if (wasteService.isWriterOfArticle(wasteId, memberPrincipal.id())) {
-            throw new WasteException(ErrorCode.OWNER_WASTE_CANT_LIKE);
-        }
     }
 }
