+++ src/test/java/freshtrash/freshtrashbackend/controller/ChatApiTest.java
@@ -4,9 +4,7 @@ import freshtrash.freshtrashbackend.Fixture.Fixture;
 import freshtrash.freshtrashbackend.config.TestSecurityConfig;
 import freshtrash.freshtrashbackend.dto.response.ChatRoomResponse;
 import freshtrash.freshtrashbackend.entity.ChatRoom;
-import freshtrash.freshtrashbackend.entity.Waste;
 import freshtrash.freshtrashbackend.service.ChatRoomService;
-import freshtrash.freshtrashbackend.service.WasteService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,12 +23,11 @@ import java.util.List;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@Import(TestSecurityConfig.class)
 @WebMvcTest(ChatApi.class)
+@Import(TestSecurityConfig.class)
 class ChatApiTest {
     @Autowired
     private MockMvc mvc;
@@ -38,12 +35,9 @@ class ChatApiTest {
     @MockBean
     private ChatRoomService chatRoomService;
 
-    @MockBean
-    private WasteService wasteService;
-
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
-    @DisplayName("채팅방 목록 조회")
     @Test
+    @DisplayName("채팅방 목록 조회")
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
     void given_memberIdAndPageable_when_getChatRooms_then_returnPagingChatRoomsResponse() throws Exception {
         // given
         Long wasteId = 1L;
@@ -82,30 +76,6 @@ class ChatApiTest {
                         .value(chatRoom.getBuyer().getNickname()))
                 .andExpect(jsonPath("$.messages.size()")
                         .value(chatRoom.getChatMessages().size()));
-        // then
-    }
-
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
-    @DisplayName("채팅 요청")
-    @Test
-    void given_wasteIdAndLoginUser_when_notSeller_then_returnChatRoomAfterGetOrCreateChatRoom() throws Exception {
-        // given
-        Long wasteId = 1L;
-        Long buyerId = 123L;
-        Long sellerId = 1L;
-        String buyerNickname = "testUser";
-        Waste waste = Fixture.createWaste();
-        ChatRoom chatRoom = Fixture.createChatRoom();
-        given(wasteService.getWaste(eq(wasteId))).willReturn(waste);
-        given(chatRoomService.getOrCreateChatRoom(eq(sellerId), eq(buyerId), eq(wasteId)))
-                .willReturn(chatRoom);
-        // when
-        mvc.perform(post("/api/v1/wastes/" + wasteId + "/chats"))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.wasteTitle").value(waste.getTitle()))
-                .andExpect(jsonPath("$.sellStatus").value("ONGOING"))
-                .andExpect(jsonPath("$.sellerNickname").value(waste.getMember().getNickname()))
-                .andExpect(jsonPath("$.buyerNickname").value(buyerNickname));
         // then
     }
 }
