+++ src/test/java/freshtrash/freshtrashbackend/controller/WasteApiTest.java
@@ -9,11 +9,13 @@ import freshtrash.freshtrashbackend.dto.request.WasteRequest;
 import freshtrash.freshtrashbackend.dto.response.WasteResponse;
 import freshtrash.freshtrashbackend.dto.security.MemberPrincipal;
 import freshtrash.freshtrashbackend.entity.Address;
+import freshtrash.freshtrashbackend.entity.ChatRoom;
 import freshtrash.freshtrashbackend.entity.Waste;
 import freshtrash.freshtrashbackend.entity.constants.SellStatus;
 import freshtrash.freshtrashbackend.entity.constants.WasteCategory;
 import freshtrash.freshtrashbackend.entity.constants.WasteStatus;
 import freshtrash.freshtrashbackend.repository.projections.FileNameSummary;
+import freshtrash.freshtrashbackend.service.ChatRoomService;
 import freshtrash.freshtrashbackend.service.LocalFileService;
 import freshtrash.freshtrashbackend.service.WasteService;
 import org.junit.jupiter.api.DisplayName;
@@ -29,6 +31,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.security.test.context.support.TestExecutionEvent;
 import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.test.web.servlet.MockMvc;
@@ -43,8 +46,8 @@ import static org.springframework.security.test.context.support.TestExecutionEve
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@Import(TestSecurityConfig.class)
 @WebMvcTest(WasteApi.class)
+@Import(TestSecurityConfig.class)
 class WasteApiTest {
     @Autowired
     private MockMvc mvc;
@@ -55,12 +58,15 @@ class WasteApiTest {
     @MockBean
     private LocalFileService localFileService;
 
+    @MockBean
+    private ChatRoomService chatRoomService;
+
     @Autowired
     private ObjectMapper objectMapper;
 
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
-    @DisplayName("폐기물 단일 조회")
     @Test
+    @DisplayName("폐기물 단일 조회")
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
     void given_wasteId_when_getWaste_then_returnWasteData() throws Exception {
         // given
         Long wasteId = 1L;
@@ -84,8 +90,8 @@ class WasteApiTest {
         // then
     }
 
-    @DisplayName("페기물 목록 조회")
     @Test
+    @DisplayName("페기물 목록 조회")
     void given_predicateAndPageable_when_getWastes_then_returnPagingWasteData() throws Exception {
         // given
         given(wasteService.getWastes(eq(null), any(Predicate.class), any(Pageable.class)))
@@ -97,24 +103,9 @@ class WasteApiTest {
         // then
     }
 
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
-    @DisplayName("관심 폐기물 목록 조회")
     @Test
-    void given_loginUserAndPageable_when_getLikedWastes_then_returnPagingWasteData() throws Exception {
-        // given
-        Long memberId = 123L;
-        given(wasteService.getLikedWastes(eq(memberId), any(Pageable.class)))
-                .willReturn(new PageImpl<>(List.of(WasteResponse.fromEntity(Fixture.createWaste()))));
-        // when
-        mvc.perform(get("/api/v1/wastes/likes"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.numberOfElements").value(1));
-        // then
-    }
-
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
     @DisplayName("폐기물 등록")
-    @Test
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
     void given_imgFileAndWasteRequest_when_requestAddWaste_then_requestValuesEqualsToReturnedWasteValues()
             throws Exception {
         Long memberId = 1L;
@@ -146,9 +137,9 @@ class WasteApiTest {
         // then
     }
 
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
-    @DisplayName("어느 하나라도 입력되지 않았을 경우 폐기물 등록 실패")
     @ParameterizedTest
+    @DisplayName("어느 하나라도 입력되지 않았을 경우 폐기물 등록 실패")
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
     @CsvSource(
             value = {
                 ", content, CLOTHING, BEST, CLOSE, 0, 12345, state, city, district, detail",
@@ -198,9 +189,9 @@ class WasteApiTest {
         // then
     }
 
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
-    @DisplayName("폐기물 수정")
     @Test
+    @DisplayName("폐기물 수정")
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
     void given_imgFileAndWasteRequest_when_requestUpdateWaste_then_requestValuesEqualsToReturnedWasteValues()
             throws Exception {
         // given
@@ -237,9 +228,9 @@ class WasteApiTest {
         // then
     }
 
-    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
-    @DisplayName("페기물 삭제")
     @Test
+    @DisplayName("페기물 삭제")
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TEST_EXECUTION)
     void given_wasteIdAndWriter_when_then_deleteWasteAndFile() throws Exception {
         // given
         Long wasteId = 1L;
@@ -251,6 +242,30 @@ class WasteApiTest {
         willDoNothing().given(wasteService).deleteWaste(eq(wasteId));
         // when
         mvc.perform(delete("/api/v1/wastes/" + wasteId)).andExpect(status().isNoContent());
+        // then
+    }
+
+    @Test
+    @DisplayName("채팅 요청")
+    @WithUserDetails(value = "testUser@gmail.com", setupBefore = TestExecutionEvent.TEST_EXECUTION)
+    void given_wasteIdAndLoginUser_when_notSeller_then_returnChatRoomAfterGetOrCreateChatRoom() throws Exception {
+        // given
+        Long wasteId = 1L;
+        Long buyerId = 123L;
+        Long sellerId = 1L;
+        String buyerNickname = "testUser";
+        Waste waste = Fixture.createWaste();
+        ChatRoom chatRoom = Fixture.createChatRoom();
+        given(wasteService.getWaste(eq(wasteId))).willReturn(waste);
+        given(chatRoomService.getOrCreateChatRoom(eq(sellerId), eq(buyerId), eq(wasteId)))
+                .willReturn(chatRoom);
+        // when
+        mvc.perform(post("/api/v1/wastes/" + wasteId + "/chats"))
+                .andExpect(status().isCreated())
+                .andExpect(jsonPath("$.wasteTitle").value(waste.getTitle()))
+                .andExpect(jsonPath("$.sellStatus").value("ONGOING"))
+                .andExpect(jsonPath("$.sellerNickname").value(waste.getMember().getNickname()))
+                .andExpect(jsonPath("$.buyerNickname").value(buyerNickname));
         // then
     }
 }
